+++ lib/components/index.ts
@@ -47,7 +47,7 @@ export { Trace } from "./primitive-components/Trace/Trace"
 export { PcbTrace } from "./primitive-components/PcbTrace"
 export { TraceHint } from "./primitive-components/TraceHint"
 export { Via } from "./primitive-components/Via"
-export { CopperPour } from "./primitive-components/CopperPour"
+export { CopperPour } from "./primitive-components/CopperPour/CopperPour"
 export { Cutout } from "./primitive-components/Cutout"
 export { Battery } from "./normal-components/Battery"
 export { PinHeader } from "./normal-components/PinHeader"
