+++ lib/components/primitive-components/CopperPour/index.ts
@@ -1,0 +1,1 @@
+export * from "./CopperPour"
