+++ tests/examples/example29-copper-pour.test.tsx
@@ -86,3 +86,35 @@ test("multiple copper pours on different layers", async () => {
 
   expect(circuit).toMatchPcbSnapshot(import.meta.path + "-multiple-layers")
 })
+
+test("copper pour should avoid traces on different nets", async () => {
+  const { circuit } = getTestFixture()
+  circuit.add(
+    <board width="20mm" height="20mm" autorouter="sequential-trace">
+      <net name="GND" />
+      <net name="VCC" />
+      <resistor
+        name="R1"
+        footprint="0805"
+        resistance="10k"
+        pcbX={-5}
+        connections={{
+          pin1: "net.VCC",
+        }}
+      />
+      <resistor
+        name="R2"
+        footprint="0805"
+        resistance="1k"
+        pcbX={5}
+        connections={{ pin2: "net.GND" }}
+      />
+      <trace from=".R1 > .pin2" to=".R2 > .pin1" />
+      <copperpour connectsTo="net.GND" layer="top" />
+    </board>,
+  )
+
+  await circuit.renderUntilSettled()
+
+  expect(circuit).toMatchPcbSnapshot(import.meta.path + "-avoid-traces")
+})
